move FileDownloader to separate file
because util.py is too long
